pyproject.toml
[tool.poetry]
name = "kindle2txt"
version = "0.1.0"
description = ""
authors = ["Shun Fukusumi <shun.fukusumi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pyautogui = "^0.9.54"
pillow = "^11.1.0"
google-genai = "^1.9.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.15.0"
flake8 = "^7.2.0"
ipython = "^9.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
